app/requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
openai>=1.40.0

firebase_admin>=6.5.0

python-docx>=1.1.2

requests>=2.32.3
